+++ src/store/street/streetSlice.ts
@@ -2,28 +2,47 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 import { createAppSlice } from "../createAppSlice";
 import { streetAPI } from "./streetAPI";
 import { Street } from "../../models/storage";
+import { getIsDemoMode } from "../../utils/config";
+import { addMockData, loadMockData, removeMockData } from "../../utils/demoUtils";
 
 export const addStreet = createAsyncThunk(
   'streets/add',
   async (street: Street) => {
-    const response = await streetAPI.addStreet(street);
-    return response.data
+    if (getIsDemoMode()) {
+      addMockData<Street>('streets', street);
+
+      return street;
+    } else {
+      const response = await streetAPI.addStreet(street);
+      return response.data;
+    }
   },
 );
 
 export const deleteStreet = createAsyncThunk(
   'streets/delete',
   async (streetId: string) => {
-    const response = await streetAPI.deleteStreet(streetId);
-    return response.data
+    if (getIsDemoMode()) {
+      removeMockData<Street>('streets', (s: Street) => { return s.id !== streetId });
+
+      return streetId;
+    } else {
+      const response = await streetAPI.deleteStreet(streetId);
+      return response.data;
+    }
   },
 );
 
 export const loadStreets = createAsyncThunk(
   'streets/fetchAll',
   async () => {
-    const response = await streetAPI.fetchStreets();
-    return response.data
+    if (getIsDemoMode()) {
+      return loadMockData<Street>('streets');
+
+    } else {
+      const response = await streetAPI.fetchStreets();
+      return response.data;
+    }
   }
 );
